chore: mark zod as external
